Move recent file handling to more appropriate place and do a little tidying.
